pyproject.toml
[project]
name = "napari-cosmx-wrapper"
version = "0.1.0"
requires-python = ">=3.8,<3.11"
dependencies = [
    "anndata>=0.9.2",
    "argparse>=1.4.0",
    "dask>=2023.5.0",
    "duckdb>=1.0.0",
    "numcodecs>=0.12.1",
    "numpy>=1.24.4,<2.0.0",
    "pandas>=2.0.3",
    "scikit-image>=0.21.0",
    "tifffile>=2023.7.10",
    "tqdm>=4.67.1",
    "vaex>=4.17.0",
    "zarr>=2.16.1",
    "napari-cosmx-fork",
    "python-dotenv>=1.0.1",
    "jupyterlab>=4.3.8",
    "polars>=1.8.2",
    "pyarrow>=17.0.0",
    "matplotlib>=3.7.5",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
gui = [
    "napari[pyqt6]",  # PyQt6 has better macOS ARM support than PyQt5
    "napari-cosmx-fork[gui]",
]

[tool.uv.sources]
napari-cosmx-fork = { workspace = true }

[tool.uv.workspace]
members = ["napari-cosmx-fork"]
